requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
faker==19.3.0

boto3==1.28.0
python-dotenv==1.0.0
Pillow==10.0.0
google-generativeai>=0.8.0

# Add this for API requests
requests==2.31.0

